+++ src/Aresta.cpp
@@ -3,8 +3,8 @@
  * @brief Implementação das funções da classe Aresta.
  */
 
-#include "Aresta.h"
-#include "No.h"
+#include "../include/Aresta.h"
+#include "../include/No.h"
 
 /**
  * @brief Construtor da classe Aresta.
