+++ src/GrafoLista.cpp
@@ -1,58 +1,45 @@
 /**
- * @file GrafoMatriz.cpp
- * @brief Implementação das funções da classe GrafoMatriz.
+ * @file GrafoLista.cpp
+ * @brief Implementação das funções da classe GrafoLista.
  */
 
-#include "GrafoMatriz.h"
+#include "../include/GrafoLista.h"
+#include "../include/Lista.h"
 #include <fstream>
 #include <iostream>
 
 using namespace std;
 
 /**
- * @brief Construtor da classe GrafoMatriz.
+ * @brief Construtor da classe GrafoLista.
  * @param ordem Número de vértices do grafo.
  * @param direcionado Indica se o grafo é direcionado (true) ou não (false).
  * @param ponderadoVertices Indica se os vértices possuem pesos (true) ou não (false).
  * @param ponderadoArestas Indica se as arestas possuem pesos (true) ou não (false).
  */
-GrafoMatriz::GrafoMatriz(int ordem, bool direcionado, bool ponderadoVertices, bool ponderadoArestas)
+GrafoLista::GrafoLista(int ordem, bool direcionado, bool ponderadoVertices, bool ponderadoArestas)
     : Grafo(ordem, direcionado, ponderadoVertices, ponderadoArestas)
 {
-    matrizAdj = new int *[ordem];
-    for (int i = 0; i < ordem; ++i)
-    {
-        matrizAdj[i] = new int[ordem];
-        for (int j = 0; j < ordem; ++j)
-        {
-            matrizAdj[i][j] = 0;
-        }
-    }
+    listaAdj = new Lista[ordem];
 }
 
 /**
- * @brief Destrutor da classe GrafoMatriz.
- * Libera a memória alocada para o grafo de matriz de adjacência.
+ * @brief Destrutor da classe GrafoLista.
+ * Libera a memória alocada para a lista de adjacência.
  */
-GrafoMatriz::~GrafoMatriz() {}
+GrafoLista::~GrafoLista()
+{
+    delete[] listaAdj;
+}
 
 /**
- * @brief Calcula o grau de um vértice no grafo.
- * O grau de um vértice é a soma dos pesos das arestas conectadas a ele.
- *
- * @param vertice O índice do vértice cujo grau será calculado.
- * @return O grau do vértice especificado.
- *
- * @warning Certifique-se de que o índice do vértice está dentro do intervalo [0, ordem-1].
+ * @brief Obtém o grau de um vértice.
+ * @param vertice Índice do vértice.
+ * @return Número de arestas conectadas ao vértice especificado.
  */
-int GrafoMatriz::getGrau(int vertice)
+int GrafoLista::getGrau(int vertice)
 {
-    int grau = 0;
-    for (int i = 0; i < ordem; i++)
-    {
-        grau += matrizAdj[vertice][i];
-    }
-    return grau;
+    return listaAdj[vertice].getTamanho();
 }
 
 /**
@@ -60,16 +47,13 @@ int GrafoMatriz::getGrau(int vertice)
  * Um grafo completo possui todas as combinações possíveis de arestas entre seus vértices.
  * @return true se o grafo é completo; caso contrário, false.
  */
-bool GrafoMatriz::ehCompleto()
+bool GrafoLista::ehCompleto()
 {
     for (int i = 0; i < ordem; i++)
     {
-        for (int j = 0; j < ordem; j++)
+        if (listaAdj[i].getTamanho() != ordem - 1)
         {
-            if (i != j && matrizAdj[i][j] == 0)
-            {
-                return false;
-            }
+            return false;
         }
     }
     return true;
@@ -80,31 +64,32 @@ bool GrafoMatriz::ehCompleto()
  * Um grafo bipartido pode ser dividido em dois subconjuntos, onde não existem arestas entre vértices do mesmo subconjunto.
  * @return true se o grafo é bipartido; caso contrário, false.
  */
-bool GrafoMatriz::ehBipartido()
+bool GrafoLista::ehBipartido()
 {
-    int *cores = new int[numVertices];
-    for (int i = 0; i < numVertices; ++i)
+    int *cores = new int[ordem];
+    for (int i = 0; i < ordem; ++i)
     {
-        cores[i] = -1;
+        cores[i] = -1; // -1 significa não colorido
     }
 
-    int *fila = new int[numVertices];
+    int *fila = new int[ordem];
     int inicio = 0, fim = 0;
 
-    cores[0] = 1;
+    cores[0] = 1; // Começa colorindo o primeiro vértice
     fila[fim++] = 0;
 
     while (inicio != fim)
     {
         int u = fila[inicio++];
-        for (int v = 0; v < numVertices; v++)
+        for (int i = 0; i < listaAdj[u].getTamanho(); ++i)
         {
-            if (matrizAdj[u][v] && cores[v] == -1)
+            int v = listaAdj[u].getElemento(i)->getIdNo();
+            if (cores[v] == -1)
             {
                 cores[v] = 1 - cores[u];
                 fila[fim++] = v;
             }
-            else if (matrizAdj[u][v] && cores[v] == cores[u])
+            else if (cores[v] == cores[u])
             {
                 delete[] cores;
                 delete[] fila;
@@ -123,7 +108,7 @@ bool GrafoMatriz::ehBipartido()
  * Um componente conexo é um subconjunto de vértices onde existe pelo menos um caminho entre cada par de vértices.
  * @return Número de componentes conexos no grafo.
  */
-int GrafoMatriz::nConexo()
+int GrafoLista::nConexo()
 {
     bool *visitado = new bool[ordem];
     for (int i = 0; i < ordem; ++i)
@@ -136,11 +121,12 @@ int GrafoMatriz::nConexo()
     auto dfs = [&](int v, auto &dfsRef) -> void
     {
         visitado[v] = true;
-        for (int i = 0; i < ordem; ++i)
+        for (int i = 0; i < listaAdj[v].getTamanho(); ++i)
         {
-            if (matrizAdj[v][i] && !visitado[i])
+            int adj = listaAdj[v].getElemento(i)->getIdNo();
+            if (!visitado[adj])
             {
-                dfsRef(i, dfsRef);
+                dfsRef(adj, dfsRef);
             }
         }
     };
@@ -163,9 +149,9 @@ int GrafoMatriz::nConexo()
  * Uma árvore é um grafo conexo sem ciclos.
  * @return true se o grafo é uma árvore; caso contrário, false.
  */
-bool GrafoMatriz::ehArvore()
+bool GrafoLista::ehArvore()
 {
-    return (nConexo() == 1 && (numVertices - 1) == getOrdem());
+    return (nConexo() == 1 && (ordem - 1) == numVertices);
 }
 
 /**
@@ -173,24 +159,62 @@ bool GrafoMatriz::ehArvore()
  * Uma ponte é uma aresta cuja remoção aumenta o número de componentes conexos do grafo.
  * @return true se existe pelo menos uma ponte; caso contrário, false.
  */
-bool GrafoMatriz::possuiPonte()
+bool GrafoLista::possuiPonte()
 {
-    for (int u = 0; u < numVertices; u++)
+    for (int u = 0; u < ordem; u++)
     {
-        for (int v = 0; v < numVertices; v++)
+        for (int i = 0; i < listaAdj[u].getTamanho(); ++i)
         {
-            if (matrizAdj[u][v])
+            int v = listaAdj[u].getElemento(i)->getIdNo();
+
+            listaAdj[u].remover(v);
+            if (!direcionado)
             {
-                matrizAdj[u][v] = 0;
-                if (nConexo() > 1)
+                listaAdj[v].remover(u);
+            }
+            bool *visitado = new bool[ordem];
+            for (int j = 0; j < ordem; ++j)
+            {
+                visitado[j] = false;
+            }
+
+            auto dfs = [&](int vertice, auto &dfsRef) -> void
+            {
+                visitado[vertice] = true;
+                for (int k = 0; k < listaAdj[vertice].getTamanho(); ++k)
                 {
-                    matrizAdj[u][v] = 1;
-                    return true;
+                    int adj = listaAdj[vertice].getElemento(k)->getIdNo();
+                    if (!visitado[adj])
+                    {
+                        dfsRef(adj, dfsRef);
+                    }
                 }
-                matrizAdj[u][v] = 1;
+            };
+
+            int componentes = 0;
+            for (int j = 0; j < ordem; j++)
+            {
+                if (!visitado[j])
+                {
+                    componentes++;
+                    dfs(j, dfs);
+                }
+            }
+
+            delete[] visitado;
+            listaAdj[u].adicionar(v);
+            if (!direcionado)
+            {
+                listaAdj[v].adicionar(u);
+            }
+
+            if (componentes > 1)
+            {
+                return true;
             }
         }
     }
+
     return false;
 }
 
@@ -199,21 +223,21 @@ bool GrafoMatriz::possuiPonte()
  * Um vértice de articulação é aquele cuja remoção aumenta o número de componentes conexos do grafo.
  * @return true se existe pelo menos um vértice de articulação; caso contrário, false.
  */
-bool GrafoMatriz::possuiArticulacao()
+bool GrafoLista::possuiArticulacao()
 {
-    bool *visitado = new bool[numVertices];
+    bool *visitado = new bool[ordem];
     int componentesOriginais = nConexo();
 
-    for (int v = 0; v < numVertices; v++)
+    for (int v = 0; v < ordem; v++)
     {
-        for (int i = 0; i < numVertices; ++i)
+        for (int i = 0; i < ordem; ++i)
         {
             visitado[i] = false;
         }
         visitado[v] = true;
 
         int componentes = 0;
-        for (int u = 0; u < numVertices; u++)
+        for (int u = 0; u < ordem; u++)
         {
             if (!visitado[u])
             {
@@ -221,11 +245,12 @@ bool GrafoMatriz::possuiArticulacao()
                 auto dfs = [&](int x, auto &dfsRef) -> void
                 {
                     visitado[x] = true;
-                    for (int i = 0; i < numVertices; i++)
+                    for (int i = 0; i < listaAdj[x].getTamanho(); ++i)
                     {
-                        if (matrizAdj[x][i] && !visitado[i])
+                        int adj = listaAdj[x].getElemento(i)->getIdNo();
+                        if (!visitado[adj])
                         {
-                            dfsRef(i, dfsRef);
+                            dfsRef(adj, dfsRef);
                         }
                     }
                 };
@@ -249,7 +274,7 @@ bool GrafoMatriz::possuiArticulacao()
  * O arquivo deve conter os dados do grafo, como número de nós, tipo de grafo e arestas.
  * @param arquivo Caminho para o arquivo de entrada.
  */
-void GrafoMatriz::carregaGrafo(const std::string &arquivo)
+void GrafoLista::carregaGrafo(const std::string &arquivo)
 {
     std::ifstream file(arquivo);
     if (!file.is_open())
@@ -277,31 +302,24 @@ void GrafoMatriz::carregaGrafo(const std::string &arquivo)
         }
     }
 
-    matrizAdj = new int *[numNos];
-    for (int i = 0; i < numNos; ++i)
-    {
-        matrizAdj[i] = new int[numNos];
-        for (int j = 0; j < numNos; ++j)
-        {
-            matrizAdj[i][j] = 0;
-        }
-    }
+    delete[] listaAdj;
+    listaAdj = new Lista[numNos];
 
     int origem, destino, peso;
     while (file >> origem >> destino >> peso)
     {
-        origem--; // Ajusta para índices baseados em 0
+        origem--;
         destino--;
-        matrizAdj[origem][destino] = peso;
+        listaAdj[origem].adicionar(destino);
         if (!direcionado)
         {
-            matrizAdj[destino][origem] = peso;
+            listaAdj[destino].adicionar(origem);
         }
         std::cout << "Aresta adicionada: " << origem + 1 << " -> " << destino + 1 << " com peso " << peso << std::endl;
     }
 
     file.close();
-    std::cout << "Matriz de Adjacência carregada com sucesso." << std::endl;
+    std::cout << "Lista de Adjacência carregada com sucesso." << std::endl;
 }
 
 /**
@@ -309,41 +327,31 @@ void GrafoMatriz::carregaGrafo(const std::string &arquivo)
  * O arquivo deve conter as informações do grafo, como tipo de estrutura, número de nós e arestas.
  * @param arquivoConfig Caminho para o arquivo de configuração.
  */
-void GrafoMatriz::novoGrafo(const std::string &arquivoConfig)
+void GrafoLista::novoGrafo(const std::string &arquivoConfig)
 {
     std::ifstream file(arquivoConfig);
     if (!file.is_open())
     {
-        std::cerr << "Erro ao abrir o arquivo de configuração: " << arquivoConfig << std::endl;
+        std::cerr << "Erro ao abrir o arquivo de configuração!" << std::endl;
         return;
     }
 
-    std::string estrutura;
-    file >> estrutura; // Lê a estrutura (matriz ou lista)
-
+    std::string tipo;
+    file >> tipo; // Ignora "matriz" ou "lista"
     file >> ordem >> direcionado >> ponderadoVertices >> ponderadoArestas;
-    matrizAdj = new int *[ordem];
-    for (int i = 0; i < ordem; ++i)
-    {
-        matrizAdj[i] = new int[ordem];
-        for (int j = 0; j < ordem; ++j)
-        {
-            matrizAdj[i][j] = 0;
-        }
-    }
+    delete[] listaAdj;
+    listaAdj = new Lista[ordem];
 
     int origem, destino, peso;
     while (file >> origem >> destino >> peso)
     {
-        origem--;
-        destino--;
-        matrizAdj[origem][destino] = peso;
+        listaAdj[origem].adicionar(destino);
         if (!direcionado)
         {
-            matrizAdj[destino][origem] = peso;
+            listaAdj[destino].adicionar(origem);
         }
     }
 
     file.close();
-    std::cout << "Novo grafo configurado como " << estrutura << "." << std::endl;
+    std::cout << "Novo grafo configurado como " << tipo << "." << std::endl;
 }
