+++ src/GrafoMatriz.cpp
@@ -1,45 +1,58 @@
 /**
- * @file GrafoLista.cpp
- * @brief Implementação das funções da classe GrafoLista.
+ * @file GrafoMatriz.cpp
+ * @brief Implementação das funções da classe GrafoMatriz.
  */
 
-#include "GrafoLista.h"
-#include "Lista.h"
+#include "../include/GrafoMatriz.h"
 #include <fstream>
 #include <iostream>
 
 using namespace std;
 
 /**
- * @brief Construtor da classe GrafoLista.
+ * @brief Construtor da classe GrafoMatriz.
  * @param ordem Número de vértices do grafo.
  * @param direcionado Indica se o grafo é direcionado (true) ou não (false).
  * @param ponderadoVertices Indica se os vértices possuem pesos (true) ou não (false).
  * @param ponderadoArestas Indica se as arestas possuem pesos (true) ou não (false).
  */
-GrafoLista::GrafoLista(int ordem, bool direcionado, bool ponderadoVertices, bool ponderadoArestas)
+GrafoMatriz::GrafoMatriz(int ordem, bool direcionado, bool ponderadoVertices, bool ponderadoArestas)
     : Grafo(ordem, direcionado, ponderadoVertices, ponderadoArestas)
 {
-    listaAdj = new Lista[ordem];
+    matrizAdj = new int *[ordem];
+    for (int i = 0; i < ordem; ++i)
+    {
+        matrizAdj[i] = new int[ordem];
+        for (int j = 0; j < ordem; ++j)
+        {
+            matrizAdj[i][j] = 0;
+        }
+    }
 }
 
 /**
- * @brief Destrutor da classe GrafoLista.
- * Libera a memória alocada para a lista de adjacência.
+ * @brief Destrutor da classe GrafoMatriz.
+ * Libera a memória alocada para o grafo de matriz de adjacência.
  */
-GrafoLista::~GrafoLista()
-{
-    delete[] listaAdj;
-}
+GrafoMatriz::~GrafoMatriz() {}
 
 /**
- * @brief Obtém o grau de um vértice.
- * @param vertice Índice do vértice.
- * @return Número de arestas conectadas ao vértice especificado.
+ * @brief Calcula o grau de um vértice no grafo.
+ * O grau de um vértice é a soma dos pesos das arestas conectadas a ele.
+ *
+ * @param vertice O índice do vértice cujo grau será calculado.
+ * @return O grau do vértice especificado.
+ *
+ * @warning Certifique-se de que o índice do vértice está dentro do intervalo [0, ordem-1].
  */
-int GrafoLista::getGrau(int vertice)
+int GrafoMatriz::getGrau(int vertice)
 {
-    return listaAdj[vertice].getTamanho();
+    int grau = 0;
+    for (int i = 0; i < ordem; i++)
+    {
+        grau += matrizAdj[vertice][i];
+    }
+    return grau;
 }
 
 /**
@@ -47,13 +60,16 @@ int GrafoLista::getGrau(int vertice)
  * Um grafo completo possui todas as combinações possíveis de arestas entre seus vértices.
  * @return true se o grafo é completo; caso contrário, false.
  */
-bool GrafoLista::ehCompleto()
+bool GrafoMatriz::ehCompleto()
 {
     for (int i = 0; i < ordem; i++)
     {
-        if (listaAdj[i].getTamanho() != ordem - 1)
+        for (int j = 0; j < ordem; j++)
         {
-            return false;
+            if (i != j && matrizAdj[i][j] == 0)
+            {
+                return false;
+            }
         }
     }
     return true;
@@ -64,32 +80,31 @@ bool GrafoLista::ehCompleto()
  * Um grafo bipartido pode ser dividido em dois subconjuntos, onde não existem arestas entre vértices do mesmo subconjunto.
  * @return true se o grafo é bipartido; caso contrário, false.
  */
-bool GrafoLista::ehBipartido()
+bool GrafoMatriz::ehBipartido()
 {
-    int *cores = new int[ordem];
-    for (int i = 0; i < ordem; ++i)
+    int *cores = new int[numVertices];
+    for (int i = 0; i < numVertices; ++i)
     {
-        cores[i] = -1; // -1 significa não colorido
+        cores[i] = -1;
     }
 
-    int *fila = new int[ordem];
+    int *fila = new int[numVertices];
     int inicio = 0, fim = 0;
 
-    cores[0] = 1; // Começa colorindo o primeiro vértice
+    cores[0] = 1;
     fila[fim++] = 0;
 
     while (inicio != fim)
     {
         int u = fila[inicio++];
-        for (int i = 0; i < listaAdj[u].getTamanho(); ++i)
+        for (int v = 0; v < numVertices; v++)
         {
-            int v = listaAdj[u].getElemento(i)->getIdNo();
-            if (cores[v] == -1)
+            if (matrizAdj[u][v] && cores[v] == -1)
             {
                 cores[v] = 1 - cores[u];
                 fila[fim++] = v;
             }
-            else if (cores[v] == cores[u])
+            else if (matrizAdj[u][v] && cores[v] == cores[u])
             {
                 delete[] cores;
                 delete[] fila;
@@ -108,7 +123,7 @@ bool GrafoLista::ehBipartido()
  * Um componente conexo é um subconjunto de vértices onde existe pelo menos um caminho entre cada par de vértices.
  * @return Número de componentes conexos no grafo.
  */
-int GrafoLista::nConexo()
+int GrafoMatriz::nConexo()
 {
     bool *visitado = new bool[ordem];
     for (int i = 0; i < ordem; ++i)
@@ -121,12 +136,11 @@ int GrafoLista::nConexo()
     auto dfs = [&](int v, auto &dfsRef) -> void
     {
         visitado[v] = true;
-        for (int i = 0; i < listaAdj[v].getTamanho(); ++i)
+        for (int i = 0; i < ordem; ++i)
         {
-            int adj = listaAdj[v].getElemento(i)->getIdNo();
-            if (!visitado[adj])
+            if (matrizAdj[v][i] && !visitado[i])
             {
-                dfsRef(adj, dfsRef);
+                dfsRef(i, dfsRef);
             }
         }
     };
@@ -149,9 +163,9 @@ int GrafoLista::nConexo()
  * Uma árvore é um grafo conexo sem ciclos.
  * @return true se o grafo é uma árvore; caso contrário, false.
  */
-bool GrafoLista::ehArvore()
+bool GrafoMatriz::ehArvore()
 {
-    return (nConexo() == 1 && (ordem - 1) == numVertices);
+    return (nConexo() == 1 && (numVertices - 1) == getOrdem());
 }
 
 /**
@@ -159,62 +173,24 @@ bool GrafoLista::ehArvore()
  * Uma ponte é uma aresta cuja remoção aumenta o número de componentes conexos do grafo.
  * @return true se existe pelo menos uma ponte; caso contrário, false.
  */
-bool GrafoLista::possuiPonte()
+bool GrafoMatriz::possuiPonte()
 {
-    for (int u = 0; u < ordem; u++)
+    for (int u = 0; u < numVertices; u++)
     {
-        for (int i = 0; i < listaAdj[u].getTamanho(); ++i)
+        for (int v = 0; v < numVertices; v++)
         {
-            int v = listaAdj[u].getElemento(i)->getIdNo();
-
-            listaAdj[u].remover(v);
-            if (!direcionado)
+            if (matrizAdj[u][v])
             {
-                listaAdj[v].remover(u);
-            }
-            bool *visitado = new bool[ordem];
-            for (int j = 0; j < ordem; ++j)
-            {
-                visitado[j] = false;
-            }
-
-            auto dfs = [&](int vertice, auto &dfsRef) -> void
-            {
-                visitado[vertice] = true;
-                for (int k = 0; k < listaAdj[vertice].getTamanho(); ++k)
+                matrizAdj[u][v] = 0;
+                if (nConexo() > 1)
                 {
-                    int adj = listaAdj[vertice].getElemento(k)->getIdNo();
-                    if (!visitado[adj])
-                    {
-                        dfsRef(adj, dfsRef);
-                    }
+                    matrizAdj[u][v] = 1;
+                    return true;
                 }
-            };
-
-            int componentes = 0;
-            for (int j = 0; j < ordem; j++)
-            {
-                if (!visitado[j])
-                {
-                    componentes++;
-                    dfs(j, dfs);
-                }
-            }
-
-            delete[] visitado;
-            listaAdj[u].adicionar(v);
-            if (!direcionado)
-            {
-                listaAdj[v].adicionar(u);
-            }
-
-            if (componentes > 1)
-            {
-                return true;
+                matrizAdj[u][v] = 1;
             }
         }
     }
-
     return false;
 }
 
@@ -223,21 +199,21 @@ bool GrafoLista::possuiPonte()
  * Um vértice de articulação é aquele cuja remoção aumenta o número de componentes conexos do grafo.
  * @return true se existe pelo menos um vértice de articulação; caso contrário, false.
  */
-bool GrafoLista::possuiArticulacao()
+bool GrafoMatriz::possuiArticulacao()
 {
-    bool *visitado = new bool[ordem];
+    bool *visitado = new bool[numVertices];
     int componentesOriginais = nConexo();
 
-    for (int v = 0; v < ordem; v++)
+    for (int v = 0; v < numVertices; v++)
     {
-        for (int i = 0; i < ordem; ++i)
+        for (int i = 0; i < numVertices; ++i)
         {
             visitado[i] = false;
         }
         visitado[v] = true;
 
         int componentes = 0;
-        for (int u = 0; u < ordem; u++)
+        for (int u = 0; u < numVertices; u++)
         {
             if (!visitado[u])
             {
@@ -245,12 +221,11 @@ bool GrafoLista::possuiArticulacao()
                 auto dfs = [&](int x, auto &dfsRef) -> void
                 {
                     visitado[x] = true;
-                    for (int i = 0; i < listaAdj[x].getTamanho(); ++i)
+                    for (int i = 0; i < numVertices; i++)
                     {
-                        int adj = listaAdj[x].getElemento(i)->getIdNo();
-                        if (!visitado[adj])
+                        if (matrizAdj[x][i] && !visitado[i])
                         {
-                            dfsRef(adj, dfsRef);
+                            dfsRef(i, dfsRef);
                         }
                     }
                 };
@@ -274,7 +249,7 @@ bool GrafoLista::possuiArticulacao()
  * O arquivo deve conter os dados do grafo, como número de nós, tipo de grafo e arestas.
  * @param arquivo Caminho para o arquivo de entrada.
  */
-void GrafoLista::carregaGrafo(const std::string &arquivo)
+void GrafoMatriz::carregaGrafo(const std::string &arquivo)
 {
     std::ifstream file(arquivo);
     if (!file.is_open())
@@ -302,24 +277,31 @@ void GrafoLista::carregaGrafo(const std::string &arquivo)
         }
     }
 
-    delete[] listaAdj;
-    listaAdj = new Lista[numNos];
+    matrizAdj = new int *[numNos];
+    for (int i = 0; i < numNos; ++i)
+    {
+        matrizAdj[i] = new int[numNos];
+        for (int j = 0; j < numNos; ++j)
+        {
+            matrizAdj[i][j] = 0;
+        }
+    }
 
     int origem, destino, peso;
     while (file >> origem >> destino >> peso)
     {
-        origem--;
+        origem--; // Ajusta para índices baseados em 0
         destino--;
-        listaAdj[origem].adicionar(destino);
+        matrizAdj[origem][destino] = peso;
         if (!direcionado)
         {
-            listaAdj[destino].adicionar(origem);
+            matrizAdj[destino][origem] = peso;
         }
         std::cout << "Aresta adicionada: " << origem + 1 << " -> " << destino + 1 << " com peso " << peso << std::endl;
     }
 
     file.close();
-    std::cout << "Lista de Adjacência carregada com sucesso." << std::endl;
+    std::cout << "Matriz de Adjacência carregada com sucesso." << std::endl;
 }
 
 /**
@@ -327,31 +309,41 @@ void GrafoLista::carregaGrafo(const std::string &arquivo)
  * O arquivo deve conter as informações do grafo, como tipo de estrutura, número de nós e arestas.
  * @param arquivoConfig Caminho para o arquivo de configuração.
  */
-void GrafoLista::novoGrafo(const std::string &arquivoConfig)
+void GrafoMatriz::novoGrafo(const std::string &arquivoConfig)
 {
     std::ifstream file(arquivoConfig);
     if (!file.is_open())
     {
-        std::cerr << "Erro ao abrir o arquivo de configuração!" << std::endl;
+        std::cerr << "Erro ao abrir o arquivo de configuração: " << arquivoConfig << std::endl;
         return;
     }
 
-    std::string tipo;
-    file >> tipo; // Ignora "matriz" ou "lista"
+    std::string estrutura;
+    file >> estrutura; // Lê a estrutura (matriz ou lista)
+
     file >> ordem >> direcionado >> ponderadoVertices >> ponderadoArestas;
-    delete[] listaAdj;
-    listaAdj = new Lista[ordem];
+    matrizAdj = new int *[ordem];
+    for (int i = 0; i < ordem; ++i)
+    {
+        matrizAdj[i] = new int[ordem];
+        for (int j = 0; j < ordem; ++j)
+        {
+            matrizAdj[i][j] = 0;
+        }
+    }
 
     int origem, destino, peso;
     while (file >> origem >> destino >> peso)
     {
-        listaAdj[origem].adicionar(destino);
+        origem--;
+        destino--;
+        matrizAdj[origem][destino] = peso;
         if (!direcionado)
         {
-            listaAdj[destino].adicionar(origem);
+            matrizAdj[destino][origem] = peso;
         }
     }
 
     file.close();
-    std::cout << "Novo grafo configurado como " << tipo << "." << std::endl;
+    std::cout << "Novo grafo configurado como " << estrutura << "." << std::endl;
 }
